Fix external links to strongloop.com (expressjs#1727) (#1728)
https links to strongloop.com are broken, and http links redirect to the
root of kyndryl.com.

This commit wraps all strongloop links with archive.org links.  The
links are set to pull the latest versions archived as of 2024-01-01 (it
looks like the site was last successfully archived at the end of 2023).